Move compose trasnform to a dedicated module
